move logger config into utils fodler
